fix : 댓글 재조합 로직 service에서 facade service로 이전, userService를 facade service에서만 의존하도록 수정
